+++ src/OpenTelemetry.Resources.Host/CHANGELOG.md
@@ -6,6 +6,10 @@
   `OTEL_DOTNET_EXPERIMENTAL_HOST_RESOURCE_ENABLE_NETWORK_ADDRESSES` is `true`.
   ([#5172](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/5172))
 
+* Updated semantic conventions to
+  [v1.44.0](https://github.com/open-telemetry/semantic-conventions/blob/v1.44.0/docs/resource/host.md).
+  ([#5172](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/5172))
+
 ## 1.18.0-beta.1
 
 Released 2026-Aug-21